Added prebuilt example for TestScene.
Same tester as in DeepSea but configured with finding DeepSea as a prebuilt
package. A small amount of duplication is needed for custom flatbuffer for
light data since installing the embedded headers could conflict when
installed on the system.

Updated Android configuration based on SDL update. Also use separate build
flavors for the different testers.